+++ dev/node/index.js
@@ -1,0 +1,2 @@
+export { default as defineReactAppConfig } from './eslint/react-app.js';
+export { prettierConfig } from './prettier.js';